[bin] Added CLI scripts to use the Datadog API to manage monitors and downtimes.

+++ bin/shared/datadog/monitor.rb
@@ -13,20 +13,21 @@ module Datadog
       @dog = Dogapi::Client.new(api_key, application_key)
     end
 
-    # Creates 10 monitors with different names and messages.
-    def create!
-      10.times do |i|
+    # Creates Monitors using the Datadog API.
+    #
+    # @param [Array[Model::Monitors]] monitors
+    def create!(monitors)
+      monitors.each do |monitor|
         options = {
           "notify_no_data" => true,
           "no_data_timeframe" => 20
         }
-        tags = ["app:webserver", "frontend"]
         dog.monitor(
-          "metric alert",
-          "avg(last_5m):sum:system.net.bytes_rcvd{host:host0} > 100",
-          name: "MONITOR #{i}",
-          message: "INDEX: #{i}",
-          tags: tags,
+          monitor.type,
+          monitor.query,
+          name: monitor.name,
+          message: monitor.message,
+          tags: monitor.tags,
           options: options
         )
       end
@@ -34,15 +35,32 @@ module Datadog
 
     # Deletes all monitors using the API.
     def delete_all!
-      get_all[1].each do |m|
+      monitors = get_all[1]
+      monitor_names = monitors.map.with_index { |m, i| "#{i + 1}: " + m["name"] }.join("\n")
+
+      UI.success("Warning! This will delete the following monitors: ")
+      UI.message(monitor_names)
+      response = UI.select("Are you sure you want to proceed?", %w[y n])
+
+      if response == "n"
+        UI.message("Exiting without deleting monitors!")
+        exit
+      end
+
+      monitors.each do |m|
         dog.delete_monitor(m["id"])
       end
+
       p dog.get_all_monitors
     end
 
-    # Returns all monitors from the APi.
+    # Returns all monitors from the API. Restricts it to only monitors created
+    # with the `github:armcburney` tag.
     def get_all
-      m = dog.get_all_monitors
+      options = {
+        "tags" => ["github:armcburney"]
+      }
+      m = dog.get_all_monitors(options: options)
       pp m
       m
     end
